requirements.txt
Django~=3.2.10
python-slugify
pillow
easy-thumbnails
django-filter
django-ckeditor
mysqlclient
django-crispy-forms
crispy_bootstrap4
django-email-verification
xhtml2pdf